Factor out hook registration during startup
