+++ module/hooks/_module.mjs
@@ -1,0 +1,4 @@
+export {default as i18n} from "./i18n.mjs";
+export {default as init} from "./init.mjs";
+export {default as ready} from "./ready.mjs";
+export {default as sidebar} from "./sidebar.mjs";
