+++ module/system/_module.mjs
@@ -1,0 +1,1 @@
+export { default as registerHooks } from './hooks.mjs';
